Revert "moved to stamper.go so stamper related code is bundled"
This reverts commit 4c1b9425e25f91f57f5f0142d739df11fb0807ca.
